feat: impl tantivy footer cache (#11895)
Design at: feat: impl tantivy footer cache

- close https://github.com/openobserve/o2-enterprise/issues/1684

### 1. Generic Bytes Cache (`infra::cache::bytes_cache::BytesCache`)

A new reusable cache for byte buffers, tagged so multiple caches can
coexist with independent metrics.

```
┌──────────────────────────────────────────┐
│              BytesCache                   │
│                                           │
│  DashMap<String, Bytes>   (readers)       │
│  Mutex<VecDeque<String>>  (cacher/FIFO)   │
│  AtomicI64                (current_memory)│
│  usize                    (max_bytes)     │
│  String                   (tag)           │
│                                           │
│  get(key) → Option<Bytes>                 │
│  put(key, value) → evict if over limit    │
│  evict() → batch-evict 5% oldest entries  │
└──────────────────────────────────────────┘
```

**Key characteristics:**
- **Concurrency:** `DashMap` for reads, `parking_lot::Mutex` around the
FIFO queue for writes.
- **Eviction:** FIFO (not LRU), evicting 5% of entries in a batch when
memory exceeds `max_bytes`. Batch eviction avoids locking on every
insertion while staying bounded.
- **Metrics:** Per-tag gauges for memory size, entry count; counters for
hits, misses, GC runs; histogram for GC time.
- **Use case flexibility:** The `tag` parameter lets different
subsystems reuse the same cache code (e.g., `"tantivy_footer_cache"`,
`"puffin_footer_cache"`).

### 2. FooterCache L2 (`FOOTER_DATA_CACHE`)

A global `BytesCache` instance wraps the `FooterCache` as an L2 cache:

```
Query
  │
  ▼
FooterCache::from_directory(dir, cache_key)
  │
  ├─ FOOTER_DATA_CACHE.get(cache_key) ──► hit → deserialize & return
  │
  └─ miss
       │
       ├─ read .footer_cache file from directory
       ├─ FOOTER_DATA_CACHE.put(cache_key, bytes)   ← store for next time
       └─ deserialize & return
```

**Serialization format** (binary, no external dependencies):
```
┌──────────────────────┐
│  raw footer data     │  ← concatenated OwnedBytes for each (file, range)
├──────────────────────┤
│  JSON metadata       │  ← FooterCacheMeta { files: Map<path, Vec<Item>> }
├──────────────────────┤
│  u64 footer_offset   │  ← LE byte offset to metadata
├──────────────────────┤
│  u32 version (1)     │  ← LE version for forward compatibility
└──────────────────────┘
```

**Configuration:**
- `ZO_INVERTED_INDEX_FOOTER_CACHE_MAX_SIZE` (env): max MB for the L2
`BytesCache`, default 5% total memory.

### 3. Puffin Footer Caching

The Puffin footer reader (`PuffinFooterBytesReader`) now also uses
`FOOTER_DATA_CACHE`. On first read, the footer tail + payload are
combined and cached under the key `"puffin_footer_{location}"`.
Subsequent reads for the same location skip the storage round-trip
entirely.

### 4. Parallel Warm-Up

`warm_up_terms` now dispatches all four warm-up categories concurrently
via `tokio::try_join!` instead of awaiting them sequentially:

```
Before (sequential):
  warm_up_fields → warm_up_fields_term → warm_up_terms → warm_up_fast_fields

After (parallel):
  ┌─ warm_up_terms ──────────┐
  ├─ warm_up_fields ─────────┤
  ├─ warm_up_fields_term ────┤
  └─ warm_up_fast_fields ────┘
           all joined
```

## Configuration Reference

| Env Variable | Default | Description |
|---|---|---|
| `ZO_INVERTED_INDEX_FOOTER_CACHE_MAX_SIZE` | 5% total memory | Max
memory for the footer L2 bytes cache |

## Metrics Reference

| Metric | Labels | Type | Description |
|---|---|---|---|
| `zo_bytes_cache_memory_size` | `tag` | Gauge | Current memory used by
a bytes cache |
| `zo_bytes_cache_entry_count` | `tag` | Gauge | Number of entries in a
bytes cache |
| `zo_bytes_cache_gc_time` | `tag` | Histogram | Time spent per GC
eviction run |
| `zo_bytes_cache_gc_count` | `tag` | Counter | Total number of GC
eviction runs |
| `zo_bytes_cache_hits_total` | `tag` | Counter | Total cache hits |
| `zo_bytes_cache_miss_total` | `tag` | Counter | Total cache misses |

+++ src/main.rs
@@ -742,8 +742,7 @@ async fn init_http_server() -> Result<(), anyhow::Error> {
             let handle = handle.clone();
             async move {
                 shutdown_signal().await;
-                handle
-                    .graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout as u64))));
+                handle.graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout))));
             }
         });
 
@@ -821,8 +820,7 @@ async fn init_http_server_without_tracing() -> Result<(), anyhow::Error> {
             let handle = handle.clone();
             async move {
                 shutdown_signal().await;
-                handle
-                    .graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout as u64))));
+                handle.graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout))));
             }
         });
